test(agent): add sysprompt abort/deny safety tests
