+++ packages/agent/test/core/policy/engine-deny-terminal.test.ts
@@ -37,72 +37,6 @@ describe("PolicyEngine deny terminal dispatch", () => {
     expect(after).toHaveBeenCalledTimes(0);
   });
 
-  it("stops system prompt composition on deny", async () => {
-    const engine = PolicyEngine.create();
-    const after = mock(
-      () =>
-        ({
-          action: "inject",
-          message: "should-not-append",
-          reason: "late",
-          policyId: "test.late",
-        }) as const,
-    );
-
-    engine.register({
-      name: "deny-context",
-      timing: "context.prepare",
-      priority: 0,
-      fn: () =>
-        ({ action: "deny", reason: "context-denied", policyId: "test.context-denied" }) as const,
-    });
-    engine.register({ name: "after", timing: "context.prepare", priority: 10, fn: after });
-
-    let thrown: unknown;
-    try {
-      await engine.dispatchSystemPrompt(baseCtx());
-    } catch (err) {
-      thrown = err;
-    }
-
-    expect(thrown).toBeInstanceOf(Error);
-    expect(String(thrown)).toContain("context-denied");
-    expect(after).toHaveBeenCalledTimes(0);
-  });
-
-  it("stops system prompt composition on abort", async () => {
-    const engine = PolicyEngine.create();
-    const after = mock(
-      () =>
-        ({
-          action: "inject",
-          message: "should-not-append",
-          reason: "late",
-          policyId: "test.late",
-        }) as const,
-    );
-
-    engine.register({
-      name: "abort-context",
-      timing: "context.prepare",
-      priority: 0,
-      fn: () =>
-        ({ action: "abort", reason: "context-aborted", policyId: "test.context-aborted" }) as const,
-    });
-    engine.register({ name: "after", timing: "context.prepare", priority: 10, fn: after });
-
-    let thrown: unknown;
-    try {
-      await engine.dispatchSystemPrompt(baseCtx());
-    } catch (err) {
-      thrown = err;
-    }
-
-    expect(thrown).toBeInstanceOf(Error);
-    expect(String(thrown)).toContain("context-aborted");
-    expect(after).toHaveBeenCalledTimes(0);
-  });
-
   it("fails closed at pre-boundary timings when policyId is missing", async () => {
     const previousNodeEnv = env().NODE_ENV;
     env().NODE_ENV = "production";
